requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.26
django-cors-headers==4.4.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
httplib2==0.31.0
idna==3.11
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-environ==0.4.54
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
